Moved Converter class in the factor package.
