+++ src/main/java/ch/idsia/crema/factor/convert/BayesianToVertex.java
@@ -1,7 +1,7 @@
 package ch.idsia.crema.factor.convert;
 
-import ch.idsia.crema.core.Converter;
 import ch.idsia.crema.core.Strides;
+import ch.idsia.crema.factor.Converter;
 import ch.idsia.crema.factor.bayesian.BayesianFactor;
 import ch.idsia.crema.factor.credal.vertex.VertexFactor;
 import com.google.common.primitives.Doubles;
