+++ src/main/java/ch/idsia/crema/factor/convert/ExtensiveVectorToExtensiveHalfspace.java
@@ -2,4 +2,7 @@ package ch.idsia.crema.factor.convert;
 
 
 public class ExtensiveVectorToExtensiveHalfspace {
+
+	// TODO
+
 }
