+++ src/main/java/ch/idsia/crema/factor/convert/IntervalToBayesianMaxEntropy.java
@@ -1,7 +1,7 @@
 package ch.idsia.crema.factor.convert;
 
-import ch.idsia.crema.core.Converter;
 import ch.idsia.crema.entropy.MaximumEntropy;
+import ch.idsia.crema.factor.Converter;
 import ch.idsia.crema.factor.bayesian.BayesianFactor;
 import ch.idsia.crema.factor.credal.linear.IntervalFactor;
 import ch.idsia.crema.utility.ArraysUtil;
