+++ src/main/java/ch/idsia/crema/factor/convert/SeparateLinearToExtensiveHalfspaceFactor.java
@@ -1,7 +1,7 @@
 package ch.idsia.crema.factor.convert;
 
-import ch.idsia.crema.core.Converter;
 import ch.idsia.crema.core.Strides;
+import ch.idsia.crema.factor.Converter;
 import ch.idsia.crema.factor.credal.linear.ExtensiveHalfspaceFactor;
 import ch.idsia.crema.factor.credal.linear.SeparateLinearFactor;
 import ch.idsia.crema.utility.IndexIterator;
